add example DARK20 solidity contract
