+++ packages/evm-development/solidity/hardhat.config.ts
@@ -1,0 +1,8 @@
+import { HardhatUserConfig } from "hardhat/config";
+import "@nomicfoundation/hardhat-toolbox";
+
+const config: HardhatUserConfig = {
+  solidity: "0.8.26",
+};
+
+export default config;
