--- packages/evm-development/source/contracts.ts/erc20.ts
@@ -1,324 +1,0 @@
-export const abi = {
-	_format: "hh-sol-artifact-1",
-	abi: [
-		{
-			inputs: [],
-			stateMutability: "nonpayable",
-			type: "constructor",
-		},
-		{
-			inputs: [
-				{
-					internalType: "address",
-					name: "spender",
-					type: "address",
-				},
-				{
-					internalType: "uint256",
-					name: "allowance",
-					type: "uint256",
-				},
-				{
-					internalType: "uint256",
-					name: "needed",
-					type: "uint256",
-				},
-			],
-			name: "ERC20InsufficientAllowance",
-			type: "error",
-		},
-		{
-			inputs: [
-				{
-					internalType: "address",
-					name: "sender",
-					type: "address",
-				},
-				{
-					internalType: "uint256",
-					name: "balance",
-					type: "uint256",
-				},
-				{
-					internalType: "uint256",
-					name: "needed",
-					type: "uint256",
-				},
-			],
-			name: "ERC20InsufficientBalance",
-			type: "error",
-		},
-		{
-			inputs: [
-				{
-					internalType: "address",
-					name: "approver",
-					type: "address",
-				},
-			],
-			name: "ERC20InvalidApprover",
-			type: "error",
-		},
-		{
-			inputs: [
-				{
-					internalType: "address",
-					name: "receiver",
-					type: "address",
-				},
-			],
-			name: "ERC20InvalidReceiver",
-			type: "error",
-		},
-		{
-			inputs: [
-				{
-					internalType: "address",
-					name: "sender",
-					type: "address",
-				},
-			],
-			name: "ERC20InvalidSender",
-			type: "error",
-		},
-		{
-			inputs: [
-				{
-					internalType: "address",
-					name: "spender",
-					type: "address",
-				},
-			],
-			name: "ERC20InvalidSpender",
-			type: "error",
-		},
-		{
-			anonymous: false,
-			inputs: [
-				{
-					indexed: true,
-					internalType: "address",
-					name: "owner",
-					type: "address",
-				},
-				{
-					indexed: true,
-					internalType: "address",
-					name: "spender",
-					type: "address",
-				},
-				{
-					indexed: false,
-					internalType: "uint256",
-					name: "value",
-					type: "uint256",
-				},
-			],
-			name: "Approval",
-			type: "event",
-		},
-		{
-			anonymous: false,
-			inputs: [
-				{
-					indexed: true,
-					internalType: "address",
-					name: "from",
-					type: "address",
-				},
-				{
-					indexed: true,
-					internalType: "address",
-					name: "to",
-					type: "address",
-				},
-				{
-					indexed: false,
-					internalType: "uint256",
-					name: "value",
-					type: "uint256",
-				},
-			],
-			name: "Transfer",
-			type: "event",
-		},
-		{
-			inputs: [
-				{
-					internalType: "address",
-					name: "owner",
-					type: "address",
-				},
-				{
-					internalType: "address",
-					name: "spender",
-					type: "address",
-				},
-			],
-			name: "allowance",
-			outputs: [
-				{
-					internalType: "uint256",
-					name: "",
-					type: "uint256",
-				},
-			],
-			stateMutability: "view",
-			type: "function",
-		},
-		{
-			inputs: [
-				{
-					internalType: "address",
-					name: "spender",
-					type: "address",
-				},
-				{
-					internalType: "uint256",
-					name: "value",
-					type: "uint256",
-				},
-			],
-			name: "approve",
-			outputs: [
-				{
-					internalType: "bool",
-					name: "",
-					type: "bool",
-				},
-			],
-			stateMutability: "nonpayable",
-			type: "function",
-		},
-		{
-			inputs: [
-				{
-					internalType: "address",
-					name: "account",
-					type: "address",
-				},
-			],
-			name: "balanceOf",
-			outputs: [
-				{
-					internalType: "uint256",
-					name: "",
-					type: "uint256",
-				},
-			],
-			stateMutability: "view",
-			type: "function",
-		},
-		{
-			inputs: [],
-			name: "decimals",
-			outputs: [
-				{
-					internalType: "uint8",
-					name: "",
-					type: "uint8",
-				},
-			],
-			stateMutability: "view",
-			type: "function",
-		},
-		{
-			inputs: [],
-			name: "name",
-			outputs: [
-				{
-					internalType: "string",
-					name: "",
-					type: "string",
-				},
-			],
-			stateMutability: "view",
-			type: "function",
-		},
-		{
-			inputs: [],
-			name: "symbol",
-			outputs: [
-				{
-					internalType: "string",
-					name: "",
-					type: "string",
-				},
-			],
-			stateMutability: "view",
-			type: "function",
-		},
-		{
-			inputs: [],
-			name: "totalSupply",
-			outputs: [
-				{
-					internalType: "uint256",
-					name: "",
-					type: "uint256",
-				},
-			],
-			stateMutability: "view",
-			type: "function",
-		},
-		{
-			inputs: [
-				{
-					internalType: "address",
-					name: "to",
-					type: "address",
-				},
-				{
-					internalType: "uint256",
-					name: "value",
-					type: "uint256",
-				},
-			],
-			name: "transfer",
-			outputs: [
-				{
-					internalType: "bool",
-					name: "",
-					type: "bool",
-				},
-			],
-			stateMutability: "nonpayable",
-			type: "function",
-		},
-		{
-			inputs: [
-				{
-					internalType: "address",
-					name: "from",
-					type: "address",
-				},
-				{
-					internalType: "address",
-					name: "to",
-					type: "address",
-				},
-				{
-					internalType: "uint256",
-					name: "value",
-					type: "uint256",
-				},
-			],
-			name: "transferFrom",
-			outputs: [
-				{
-					internalType: "bool",
-					name: "",
-					type: "bool",
-				},
-			],
-			stateMutability: "nonpayable",
-			type: "function",
-		},
-	],
-	bytecode:
-		"0x60806040523480156200001157600080fd5b506040518060400160405280600d81526020017f4d61696e7361696c4552433230000000000000000000000000000000000000008152506040518060400160405280600d81526020017f4d41494e5341494c45524332300000000000000000000000000000000000000081525081600390816200008f9190620005fe565b508060049081620000a19190620005fe565b505050620000c1336a52b7d2dcc80cd2e4000000620000c760201b60201c565b6200081c565b600073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff16036200013c5760006040517fec442f050000000000000000000000000000000000000000000000000000000081526004016200013391906200072a565b60405180910390fd5b62000150600083836200015460201b60201c565b5050565b600073ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff1603620001aa5780600260008282546200019d919062000776565b9250508190555062000280565b60008060008573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020016000205490508181101562000239578381836040517fe450d38c0000000000000000000000000000000000000000000000000000000081526004016200023093929190620007c2565b60405180910390fd5b8181036000808673ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002081905550505b600073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff1603620002cb578060026000828254039250508190555062000318565b806000808473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020600082825401925050819055505b8173ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef83604051620003779190620007ff565b60405180910390a3505050565b600081519050919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052604160045260246000fd5b7f4e487b7100000000000000000000000000000000000000000000000000000000600052602260045260246000fd5b600060028204905060018216806200040657607f821691505b6020821081036200041c576200041b620003be565b5b50919050565b60008190508160005260206000209050919050565b60006020601f8301049050919050565b600082821b905092915050565b600060088302620004867fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff8262000447565b62000492868362000447565b95508019841693508086168417925050509392505050565b6000819050919050565b6000819050919050565b6000620004df620004d9620004d384620004aa565b620004b4565b620004aa565b9050919050565b6000819050919050565b620004fb83620004be565b620005136200050a82620004e6565b84845462000454565b825550505050565b600090565b6200052a6200051b565b62000537818484620004f0565b505050565b5b818110156200055f576200055360008262000520565b6001810190506200053d565b5050565b601f821115620005ae57620005788162000422565b620005838462000437565b8101602085101562000593578190505b620005ab620005a28562000437565b8301826200053c565b50505b505050565b600082821c905092915050565b6000620005d360001984600802620005b3565b1980831691505092915050565b6000620005ee8383620005c0565b9150826002028217905092915050565b620006098262000384565b67ffffffffffffffff8111156200062557620006246200038f565b5b620006318254620003ed565b6200063e82828562000563565b600060209050601f83116001811462000676576000841562000661578287015190505b6200066d8582620005e0565b865550620006dd565b601f198416620006868662000422565b60005b82811015620006b05784890151825560018201915060208501945060208101905062000689565b86831015620006d05784890151620006cc601f891682620005c0565b8355505b6001600288020188555050505b505050505050565b600073ffffffffffffffffffffffffffffffffffffffff82169050919050565b60006200071282620006e5565b9050919050565b620007248162000705565b82525050565b600060208201905062000741600083018462000719565b92915050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052601160045260246000fd5b60006200078382620004aa565b91506200079083620004aa565b9250828201905080821115620007ab57620007aa62000747565b5b92915050565b620007bc81620004aa565b82525050565b6000606082019050620007d9600083018662000719565b620007e86020830185620007b1565b620007f76040830184620007b1565b949350505050565b6000602082019050620008166000830184620007b1565b92915050565b610e55806200082c6000396000f3fe608060405234801561001057600080fd5b50600436106100935760003560e01c8063313ce56711610066578063313ce5671461013457806370a082311461015257806395d89b4114610182578063a9059cbb146101a0578063dd62ed3e146101d057610093565b806306fdde0314610098578063095ea7b3146100b657806318160ddd146100e657806323b872dd14610104575b600080fd5b6100a0610200565b6040516100ad9190610aa9565b60405180910390f35b6100d060048036038101906100cb9190610b64565b610292565b6040516100dd9190610bbf565b60405180910390f35b6100ee6102b5565b6040516100fb9190610be9565b60405180910390f35b61011e60048036038101906101199190610c04565b6102bf565b60405161012b9190610bbf565b60405180910390f35b61013c6102ee565b6040516101499190610c73565b60405180910390f35b61016c60048036038101906101679190610c8e565b6102f7565b6040516101799190610be9565b60405180910390f35b61018a61033f565b6040516101979190610aa9565b60405180910390f35b6101ba60048036038101906101b59190610b64565b6103d1565b6040516101c79190610bbf565b60405180910390f35b6101ea60048036038101906101e59190610cbb565b6103f4565b6040516101f79190610be9565b60405180910390f35b60606003805461020f90610d2a565b80601f016020809104026020016040519081016040528092919081815260200182805461023b90610d2a565b80156102885780601f1061025d57610100808354040283529160200191610288565b820191906000526020600020905b81548152906001019060200180831161026b57829003601f168201915b5050505050905090565b60008061029d61047b565b90506102aa818585610483565b600191505092915050565b6000600254905090565b6000806102ca61047b565b90506102d7858285610495565b6102e2858585610529565b60019150509392505050565b60006012905090565b60008060008373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020549050919050565b60606004805461034e90610d2a565b80601f016020809104026020016040519081016040528092919081815260200182805461037a90610d2a565b80156103c75780601f1061039c576101008083540402835291602001916103c7565b820191906000526020600020905b8154815290600101906020018083116103aa57829003601f168201915b5050505050905090565b6000806103dc61047b565b90506103e9818585610529565b600191505092915050565b6000600160008473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060008373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002054905092915050565b600033905090565b610490838383600161061d565b505050565b60006104a184846103f4565b90507fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff81146105235781811015610513578281836040517ffb8f41b200000000000000000000000000000000000000000000000000000000815260040161050a93929190610d6a565b60405180910390fd5b6105228484848403600061061d565b5b50505050565b600073ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff160361059b5760006040517f96c6fd1e0000000000000000000000000000000000000000000000000000000081526004016105929190610da1565b60405180910390fd5b600073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff160361060d5760006040517fec442f050000000000000000000000000000000000000000000000000000000081526004016106049190610da1565b60405180910390fd5b6106188383836107f4565b505050565b600073ffffffffffffffffffffffffffffffffffffffff168473ffffffffffffffffffffffffffffffffffffffff160361068f5760006040517fe602df050000000000000000000000000000000000000000000000000000000081526004016106869190610da1565b60405180910390fd5b600073ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff16036107015760006040517f94280d620000000000000000000000000000000000000000000000000000000081526004016106f89190610da1565b60405180910390fd5b81600160008673ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060008573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020016000208190555080156107ee578273ffffffffffffffffffffffffffffffffffffffff168473ffffffffffffffffffffffffffffffffffffffff167f8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925846040516107e59190610be9565b60405180910390a35b50505050565b600073ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff160361084657806002600082825461083a9190610deb565b92505081905550610919565b60008060008573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020549050818110156108d2578381836040517fe450d38c0000000000000000000000000000000000000000000000000000000081526004016108c993929190610d6a565b60405180910390fd5b8181036000808673ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002081905550505b600073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff160361096257806002600082825403925050819055506109af565b806000808473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020600082825401925050819055505b8173ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef83604051610a0c9190610be9565b60405180910390a3505050565b600081519050919050565b600082825260208201905092915050565b60005b83811015610a53578082015181840152602081019050610a38565b60008484015250505050565b6000601f19601f8301169050919050565b6000610a7b82610a19565b610a858185610a24565b9350610a95818560208601610a35565b610a9e81610a5f565b840191505092915050565b60006020820190508181036000830152610ac38184610a70565b905092915050565b600080fd5b600073ffffffffffffffffffffffffffffffffffffffff82169050919050565b6000610afb82610ad0565b9050919050565b610b0b81610af0565b8114610b1657600080fd5b50565b600081359050610b2881610b02565b92915050565b6000819050919050565b610b4181610b2e565b8114610b4c57600080fd5b50565b600081359050610b5e81610b38565b92915050565b60008060408385031215610b7b57610b7a610acb565b5b6000610b8985828601610b19565b9250506020610b9a85828601610b4f565b9150509250929050565b60008115159050919050565b610bb981610ba4565b82525050565b6000602082019050610bd46000830184610bb0565b92915050565b610be381610b2e565b82525050565b6000602082019050610bfe6000830184610bda565b92915050565b600080600060608486031215610c1d57610c1c610acb565b5b6000610c2b86828701610b19565b9350506020610c3c86828701610b19565b9250506040610c4d86828701610b4f565b9150509250925092565b600060ff82169050919050565b610c6d81610c57565b82525050565b6000602082019050610c886000830184610c64565b92915050565b600060208284031215610ca457610ca3610acb565b5b6000610cb284828501610b19565b91505092915050565b60008060408385031215610cd257610cd1610acb565b5b6000610ce085828601610b19565b9250506020610cf185828601610b19565b9150509250929050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052602260045260246000fd5b60006002820490506001821680610d4257607f821691505b602082108103610d5557610d54610cfb565b5b50919050565b610d6481610af0565b82525050565b6000606082019050610d7f6000830186610d5b565b610d8c6020830185610bda565b610d996040830184610bda565b949350505050565b6000602082019050610db66000830184610d5b565b92915050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052601160045260246000fd5b6000610df682610b2e565b9150610e0183610b2e565b9250828201905080821115610e1957610e18610dbc565b5b9291505056fea264697066735822122015f3a37bdbb48457206f9cc74341926c0f2929afa9ee53693f5890d447229a0264736f6c63430008140033",
-	contractName: "MainsailERC20",
-	deployedLinkReferences: {},
-	linkReferences: {},
-	sourceName: "contracts/MainsailERC20.sol",
-};
