+++ packages/evm-development/source/contracts.ts/index.ts
@@ -1,1 +1,1 @@
-export * as ERC20 from "./erc20.js";
+export * as ERC20 from "./dark20.js";
