+++ pnpm-workspace.yaml
@@ -2,3 +2,4 @@ packages:
     - "packages/**"
     - "tests/**"
     - "!tests/e2e/**"
+    - "!packages/evm-development/solidity"
